backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
supabase==2.15.0
selenium==4.22.0
